+++ x/transferinject/types/expected_keepers.go
@@ -3,6 +3,7 @@ package types
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/bank/types"
+	delayedacktypes "github.com/dymensionxyz/dymension/v3/x/delayedack/types"
 
 	rollapptypes "github.com/dymensionxyz/dymension/v3/x/rollapp/types"
 )
@@ -14,9 +15,9 @@ type BankKeeper interface {
 
 type RollappKeeper interface {
 	SetRollapp(ctx sdk.Context, rollapp rollapptypes.Rollapp)
-	ExtractRollappFromChannel(
+	GetValidTransfer(
 		ctx sdk.Context,
-		rollappPortOnHub string,
-		rollappChannelOnHub string,
-	) (*rollapptypes.Rollapp, error)
+		packetData []byte,
+		raPortOnHub, raChanOnHub string,
+	) (data delayedacktypes.TransferData, err error)
 }
